fix: add missing CRDs for Gateway API, Knative Serving, and Knative Eventing (#72)
The Helm chart was failing to install because it was trying to create
resources (Gateway, GatewayClass, Certificate, Image) that depend on
CRDs that were not present in the crds/ directory.

This commit adds the following CRDs:
- Gateway API v1.3.0 CRDs (GatewayClass, Gateway, HTTPRoute, etc.)
- Knative Serving v1.16.0 CRDs (Certificate, Image, Service, etc.)
- Knative Eventing v1.16.0 CRDs (Broker, Trigger, etc.)

These CRDs are installed by Helm before the templates are processed,
ensuring that all dependent resources can be created successfully.